+++ GeoAdsPlatform/Models/AdInfo.cs
@@ -9,8 +9,8 @@ namespace GeoAdsPlatform.Models
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public float Lat { get; set; }
-        public float Lon { get; set; }
+        public double Lat { get; set; }
+        public double Lon { get; set; }
         public int Radius { get; set; }
     }
